style(suite): consolidate colors
